# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Contatos",
        "description": "Esta é a documentação da API contatos.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/contacts": {
            "get": {
                "tags": [
                    "Contatos"
                ],
                "summary": "Lista todos os contatos",
                "operationId": "58c0129b81e10aa54478936abda0eb0b",
                "responses": {
                    "200": {
                        "description": "Lista de contatos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Contact"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro ao processar a requisição",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Erro ao listar contatos"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Contatos"
                ],
                "summary": "Cria um novo contato",
                "operationId": "e70338b5beb975934484a7223cd45163",
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Contato criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Contact"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro ao processar a requisição",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Erro ao criar contato"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/contacts/{contact}": {
            "get": {
                "tags": [
                    "Contatos"
                ],
                "summary": "Obtém detalhes de um contato",
                "operationId": "48f3b6b6be30027138883726b758f5c3",
                "parameters": [
                    {
                        "name": "contact",
                        "in": "path",
                        "description": "ID do contato",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do contato",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro ao processar a requisição",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Erro ao obter contato"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Contatos"
                ],
                "summary": "Atualiza um contato",
                "operationId": "aaeb8286b242c3ea0acf1e1cc43d5a48",
                "parameters": [
                    {
                        "name": "contact",
                        "in": "path",
                        "description": "ID do contato",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Contato atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Contact"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro ao processar a requisição",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Erro ao atualizar contato"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Contatos"
                ],
                "summary": "Exclui um contato",
                "operationId": "9a033d4e695e34378889a91f795dbcec",
                "parameters": [
                    {
                        "name": "contact",
                        "in": "path",
                        "description": "ID do contato",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Contato excluído com sucesso"
                    },
                    "400": {
                        "description": "Erro ao processar a requisição",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Erro ao excluir contato"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Contact": {
                "required": [
                    "id",
                    "type",
                    "contact"
                ],
                "properties": {
                    "id": {
                        "description": "ID do contato",
                        "type": "integer"
                    },
                    "type": {
                        "description": "Tipo de contato",
                        "type": "string"
                    },
                    "contact": {
                        "description": "Informação do contato",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}